Commit Changes for bug 533
